+++ frontend/src/features/editor/dropdowns/index.tsx
@@ -1,0 +1,4 @@
+import { FontSizeDropDown } from './fontSizeDropdown';
+import { FontStyleDropDown } from './fontStyleDropdown';
+
+export { FontSizeDropDown, FontStyleDropDown };
