+++ frontend/src/features/editor/plugins/index.tsx
@@ -1,0 +1,4 @@
+import { MyBoldPlugin } from './myBoldPlugin';
+import { MyHeadingPlugin } from './myHeadingPlugin';
+import { MyCustomAutoFocusPlugin } from './myCustomAutoFocusPlugin';
+export { MyBoldPlugin, MyHeadingPlugin, MyCustomAutoFocusPlugin };
